feat: move settings provider near the component
