+++ .github/workflows/build.yml
@@ -2,8 +2,6 @@ name: 'Publish'
 
 on:
   push:
-    branches:
-      - "feat/updater"
     tags:
       - "v*"
