+++ src-tauri/tauri.conf.json
@@ -1,6 +1,6 @@
 {
   "productName": "mytime",
-  "version": "0.8.5",
+  "version": "0.8.4",
   "identifier": "com.farpi.mytime",
   "build": {
     "beforeDevCommand": "npm run dev",
