+++ src/components/Settings/Provider.tsx
@@ -1,9 +1,7 @@
 import { createContext, ReactNode, useState, useEffect } from "react";
-import { Setting, ViewType } from "../hooks/useSettings";
-import useSettings from "../hooks/useSettings";
-import Settings from "../components/Settings/Settings";
+import useSettings, { Setting, ViewType } from "../../hooks/useSettings";
+import Settings from "./Settings";
 
-// TODO: move into the settings folder?
 const SettingsContext = createContext<{
   setting: Setting | null;
   isIntegrationValid: boolean;
